+++ engine/select_members.py
@@ -157,7 +157,6 @@ def check_selection_by_ids(
     logging.getLogger().setLevel(logging.ERROR)
 
     variables = set()
-    rules = ""
 
     for mem in member_ids:
         m_id = str(mem) if not member_type else member_type + "_" + str(mem)
@@ -167,7 +166,6 @@ def check_selection_by_ids(
             FileInfo(stats_file_name.format(member_id="ref")),
             FileInfo(stats_file_name.format(member_id=m_id)),
             factor,
-            rules,
         )
 
         if out:
