+++ tests/engine/test_check.py
@@ -3,206 +3,196 @@ This module contains test cases to validate the functionality
 of check CLI commands.
 """
 
+import os
+
 import numpy as np
 import pandas as pd
 import pytest
+from click.testing import CliRunner
 
-from util.constants import CHECK_THRESHOLD
-from util.dataframe_ops import (
-    check_variable,
-    compute_rel_diff_dataframe,
-    split_feedback_dataset,
-)
+from engine.check import check
 
 
 @pytest.fixture(name="stats_dataframes", scope="function")
-def fixture_dataframes():
+def fixture_dataframes(tmp_dir):
     """
     Create stats dataframes and reference tolerances.
     """
+
+    file_ids = ["NetCDF:*atm_3d*.nc"]
+    variables = ["var_1", "var_2"]
+    times = range(4)
+    stats = ["max", "mean", "min"]
+
     index = pd.MultiIndex.from_arrays(
-        [
-            ["NetCDF:*atm_3d*.nc"] * 4,
-            ["var_1"] * 3 + ["var_2"],
-            list(range(3)) + [0],
-        ],
+        [file_ids * 4, ["var_1"] * 3 + ["var_2"], list(range(3)) + [0]],
         names=["file_ID", "variable", "height"],
     )
-    columns = pd.MultiIndex.from_product(
-        [range(4), ["max", "mean", "min"]], names=["time", "statistic"]
-    )
+    columns = pd.MultiIndex.from_product([times, stats], names=["time", "statistic"])
 
     array1 = np.linspace(0.9, 1.9, 4 * 12).reshape(4, 12)
     array1[:2] *= -1
     df1 = pd.DataFrame(array1, index=index, columns=columns)
 
-    array2 = np.linspace(1.1, 2.1, 4 * 12).transpose().reshape(4, 12)
+    array2 = np.linspace(1.1, 2.1, 4 * 12).reshape(4, 12)
     array2[:2] *= -1
     df2 = pd.DataFrame(array2, index=index, columns=columns)
 
-    tol_large = pd.DataFrame(
-        np.ones((2, 12)) * 0.21, index=["var_1", "var_2"], columns=columns
+    base_index = pd.Index(variables, name="variable")
+    tol_large_base = pd.DataFrame(
+        np.ones((2, 12)) * 0.21, index=base_index, columns=columns
     )
-    tol_small = pd.DataFrame(
-        np.ones((2, 12)) * 0.06, index=["var_1", "var_2"], columns=columns
+    tol_small_base = pd.DataFrame(
+        np.ones((2, 12)) * 0.06, index=base_index, columns=columns
     )
 
-    return df1, df2, tol_large, tol_small
+    tol_large = pd.concat(
+        [tol_large_base] * len(file_ids), keys=file_ids, names=["file_ID", "variable"]
+    )
+    tol_small = pd.concat(
+        [tol_small_base] * len(file_ids), keys=file_ids, names=["file_ID", "variable"]
+    )
+
+    df1_stats = os.path.join(tmp_dir, "stats1.csv")
+    df2_stats = os.path.join(tmp_dir, "stats2.csv")
+    tol_large_stats = os.path.join(tmp_dir, "tol_large.csv")
+    tol_small_stats = os.path.join(tmp_dir, "tol_small.csv")
+
+    df1.to_csv(df1_stats)
+    df2.to_csv(df2_stats)
+    tol_large.to_csv(tol_large_stats)
+    tol_small.to_csv(tol_small_stats)
+
+    yield df1_stats, df2_stats, tol_large_stats, tol_small_stats
 
 
 @pytest.fixture(name="fof_datasets", scope="function")
-def fixture_fof_datasets(sample_dataset_fof):
+def fixture_fof_datasets(tmp_dir, sample_dataset_fof):
     """
     Create fof dataset and reference tolerances.
     """
+
     ds1 = sample_dataset_fof
     ds2 = ds1.copy(deep=True)
-
     ds2["veri_data"] = (("d_body",), ds2["veri_data"].values * 1.55)
-    print(ds1["veri_data"])
-    print(ds2["veri_data"])
 
     n_body_size = ds1.sizes["d_body"]
 
     tol_large = pd.DataFrame({"veri_data": np.full(n_body_size, 5)})
     tol_small = pd.DataFrame({"veri_data": np.full(n_body_size, 0.06)})
 
-    return ds1, ds2, tol_large, tol_small
+    ds1_file = os.path.join(tmp_dir, "fof1.nc")
+    ds2_file = os.path.join(tmp_dir, "fof2.nc")
+    tol_large_file = os.path.join(tmp_dir, "fof_tol_large.csv")
+    tol_small_file = os.path.join(tmp_dir, "fof_tol_small.csv")
+
+    ds1.to_netcdf(ds1_file)
+    ds2.to_netcdf(ds2_file)
+    tol_large.to_csv(tol_large_file)
+    tol_small.to_csv(tol_small_file)
+
+    yield ds1_file, ds2_file, tol_large_file, tol_small_file
 
 
-def _check(df1, df2, tol_large, tol_small, file_type):
+def test_check_cli_stats(stats_dataframes):
+    """
+    Test that is not validated in the case of large tolerances but
+    is validated in the case of small tolerances.
+    """
 
-    diff_df = compute_rel_diff_dataframe(df1, df2)
+    df1_stats, df2_stats, tol_large, tol_small = stats_dataframes
 
-    if file_type == "stats":
-        diff_df = diff_df.groupby(["variable"]).max()
-    if file_type == "fof":
-        diff_df = diff_df.to_frame()
-
-    out1, err1, _ = check_variable(diff_df, tol_large)
-    out2, err2, _ = check_variable(diff_df, tol_small)
-
-    assert out1, (
-        "Check with large tolerances did not validate! "
-        + f"Here is the DataFrame:\n{err1}"
-    )
-    assert not out2, (
-        f"Check with small tolerances did validate! " f"Here is the DataFrame:\n{err2}"
-    )
-
-
-def test_check_stats(stats_dataframes):
-    df1, df2, tol_large, tol_small = stats_dataframes
-    _check(df1, df2, tol_large, tol_small, file_type="stats")
-
-
-def test_check_fof(fof_datasets):
-    ds1, ds2, tol_large, tol_small = fof_datasets
-    _, _, ds_veri1 = split_feedback_dataset(ds1)
-    _, _, ds_veri2 = split_feedback_dataset(ds2)
-    df_veri1 = ds_veri1.to_dataframe().reset_index()
-    df_veri2 = ds_veri2.to_dataframe().reset_index()
-    _check(
-        df_veri1["veri_data"],
-        df_veri2["veri_data"],
-        tol_large,
-        tol_small,
-        file_type="fof",
+    runner = CliRunner()
+    result = runner.invoke(
+        check,
+        [
+            "--reference-files",
+            df1_stats,
+            "--current-files",
+            df2_stats,
+            "--tolerance-files",
+            tol_small,
+            "--factor",
+            "1.0",
+            "--fof-types",
+            "",
+        ],
     )
 
+    assert result.exit_code == 1
 
-def test_check_one_zero_stats(stats_dataframes):
-    """
-    Test that a null value in ds1 causes failure,
-    and that a variation within tolerance is accepted.
-    """
-    df1, df2, tol_large, tol_small = stats_dataframes
-    df1 = df1.copy()
-    df1.loc[("NetCDF:*atm_3d*.nc", "var_1", 2), (0, "max")] = 0
-
-    diff_df = compute_rel_diff_dataframe(df1, df2)
-    diff_df = diff_df.groupby(["variable"]).max()
-    out, err, _ = check_variable(diff_df, tol_large)
-
-    assert not out, f"Check with 0-value reference validated incorrectly:\n{err}"
-
-    df2 = df2.copy()
-    df2.loc[("NetCDF:*atm_3d*.nc", "var_1", 2), (0, "max")] = CHECK_THRESHOLD / 2
-    _check(df1, df2, tol_large, tol_small, file_type="stats")
-
-
-def test_check_one_zero_fof(fof_datasets):
-    """
-    Similar to test_check_one_zero, but applied to fof files.
-    """
-    ds1, ds2, tol_large, tol_small = fof_datasets
-
-    ds2_copy = ds2.copy(deep=True)
-    ds1["veri_data"][2] = 0
-    ds2_copy["veri_data"][2] = CHECK_THRESHOLD / 2
-
-    _, _, ds_veri1 = split_feedback_dataset(ds1)
-    _, _, ds_veri2 = split_feedback_dataset(ds2)
-    df_veri1 = ds_veri1.to_dataframe().reset_index()
-    df_veri2 = ds_veri2.to_dataframe().reset_index()
-
-    diff_df = compute_rel_diff_dataframe(df_veri1["veri_data"], df_veri2["veri_data"])
-    diff_df = diff_df.to_frame()
-
-    out, err, _ = check_variable(diff_df, tol_large)
-
-    assert not out, f"Check with 0-value reference validated incorrectly:\n{err}"
-
-    _, _, ds_veri2_copy = split_feedback_dataset(ds2_copy)
-    ds_veri2_copy = ds_veri2_copy.copy(deep=True)
-    ds_veri2_copy["veri_data"][2] = CHECK_THRESHOLD / 2
-    _check(
-        df_veri1["veri_data"],
-        ds_veri2_copy["veri_data"],
-        tol_large,
-        tol_small,
-        file_type="fof",
+    runner = CliRunner()
+    result = runner.invoke(
+        check,
+        [
+            "--reference-files",
+            df1_stats,
+            "--current-files",
+            df2_stats,
+            "--tolerance-files",
+            tol_large,
+            "--factor",
+            "1.0",
+            "--fof-types",
+            "",
+        ],
     )
 
+    assert result.exit_code == 0
 
-def test_check_smalls_stats(stats_dataframes):
+
+def test_check_cli_fof(fof_datasets):
     """
-    Both values are close to 0 and should be accepted even though
-    their relative difference is large.
-    Close to 0 means < util.constants.CHECK_THRESHOLD.
+    Same as before but for fof files.
     """
 
-    df1, df2, tol_large, tol_small = stats_dataframes
-    df1 = df1.copy()
-    df2 = df2.copy()
+    df1, df2, tol_large, tol_small = fof_datasets
 
-    df1.loc["var_1", (2, "min")] = CHECK_THRESHOLD * 1e-5
-    df2.loc["var_1", (2, "min")] = -CHECK_THRESHOLD / 2
+    rules = {
+        "check": [13, 18, 32],
+        "state": [1, 5, 7, 9],
+        "r_check": [13, 18, 32],
+        "r_state": [1, 5, 7, 9],
+    }
 
-    _check(df1, df2, tol_large, tol_small, file_type="stats")
-
-
-def test_check_smalls_fof(fof_datasets):
-    """
-    Test similar to test_check_smalls but on FOF datasets.
-    """
-    ds1, ds2, tol_large, tol_small = fof_datasets
-
-    ds1 = ds1.copy(deep=True)
-    ds2 = ds2.copy(deep=True)
-
-    ds1["veri_data"][2] = CHECK_THRESHOLD * 1e-5
-    ds2["veri_data"][2] = -CHECK_THRESHOLD / 2
-
-    _, _, ds_veri1 = split_feedback_dataset(ds1)
-    _, _, ds_veri2 = split_feedback_dataset(ds2)
-    df_veri1 = ds_veri1.to_dataframe().reset_index()
-    df_veri2 = ds_veri2.to_dataframe().reset_index()
-
-    _check(
-        df_veri1["veri_data"],
-        df_veri2["veri_data"],
-        tol_large,
-        tol_small,
-        file_type="fof",
+    runner = CliRunner()
+    result = runner.invoke(
+        check,
+        [
+            "--reference-files",
+            df1,
+            "--current-files",
+            df2,
+            "--tolerance-files",
+            tol_small,
+            "--factor",
+            "1.0",
+            "--fof-types",
+            "",
+            "--rules",
+            rules,
+        ],
     )
+
+    assert result.exit_code == 1
+
+    runner = CliRunner()
+    result = runner.invoke(
+        check,
+        [
+            "--reference-files",
+            df1,
+            "--current-files",
+            df2,
+            "--tolerance-files",
+            tol_large,
+            "--factor",
+            "1.0",
+            "--fof-types",
+            "",
+            "--rules",
+            rules,
+        ],
+    )
+
+    assert result.exit_code == 0
